Allow legacy uuid string on cancel/reject/expire/query order
