+++ src/contract.rs
@@ -4018,3841 +4018,6 @@ mod tests {
     }
 
     #[test]
-    fn cancel_ask_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // cancel ask order
-        let asker_info = mock_info("asker", &[]);
-
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let cancel_ask_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            asker_info.clone(),
-            cancel_ask_msg,
-        );
-
-        match cancel_ask_response {
-            Ok(cancel_ask_response) => {
-                assert_eq!(cancel_ask_response.attributes.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.attributes[0],
-                    attr("action", "cancel_ask")
-                );
-                assert_eq!(
-                    cancel_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(cancel_ask_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: asker_info.sender.to_string(),
-                        amount: coins(100, "base_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_ask_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let test_marker: Marker = from_binary(&Binary::from(marker_json)).unwrap();
-        deps.querier.with_markers(vec![test_marker]);
-
-        // create bid data
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("asker"),
-                class: AskOrderClass::Basic,
-                base: "base_1".into(),
-                quote: "quote_1".into(),
-                price: "2".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // cancel ask order
-        let asker_info = mock_info("asker", &[]);
-
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_ask_response = execute(deps.as_mut(), mock_env(), asker_info, cancel_ask_msg);
-
-        match cancel_ask_response {
-            Ok(cancel_ask_response) => {
-                assert_eq!(cancel_ask_response.attributes.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.attributes[0],
-                    attr("action", "cancel_ask")
-                );
-                assert_eq!(
-                    cancel_ask_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-
-                assert_eq!(cancel_ask_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_ask_response.messages[0].msg,
-                    transfer_marker_coins(
-                        100,
-                        "base_1",
-                        Addr::unchecked("asker"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_ask_convertible_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("approver_1"), Addr::unchecked("approver_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "con_base_1".into(),
-                class: AskOrderClass::Convertible {
-                    status: AskOrderStatus::Ready {
-                        approver: Addr::unchecked("approver_1"),
-                        converted_base: coin(100, "base_denom"),
-                    },
-                },
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // cancel ask order
-        let asker_info = mock_info("asker", &[]);
-
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let cancel_ask_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            asker_info.clone(),
-            cancel_ask_msg,
-        );
-
-        match cancel_ask_response {
-            Ok(cancel_ask_response) => {
-                assert_eq!(cancel_ask_response.attributes.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.attributes[0],
-                    attr("action", "cancel_ask")
-                );
-                assert_eq!(
-                    cancel_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(cancel_ask_response.messages.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: asker_info.sender.to_string(),
-                        amount: coins(100, "con_base_1"),
-                    })
-                );
-                assert_eq!(
-                    cancel_ask_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "approver_1".to_string(),
-                        amount: coins(100, "base_denom"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_ask_convertible_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("approver_1"), Addr::unchecked("approver_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let convertible_marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"con_base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"con_base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let base_marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let base_marker: Marker = from_binary(&Binary::from(base_marker_json)).unwrap();
-        let convertible_marker: Marker =
-            from_binary(&Binary::from(convertible_marker_json)).unwrap();
-        deps.querier
-            .with_markers(vec![base_marker, convertible_marker]);
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "con_base_1".into(),
-                class: AskOrderClass::Convertible {
-                    status: AskOrderStatus::Ready {
-                        approver: Addr::unchecked("approver_1"),
-                        converted_base: coin(100, "base_1"),
-                    },
-                },
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // cancel ask order
-        let asker_info = mock_info("asker", &[]);
-
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let cancel_ask_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            asker_info.clone(),
-            cancel_ask_msg,
-        );
-
-        match cancel_ask_response {
-            Ok(cancel_ask_response) => {
-                assert_eq!(cancel_ask_response.attributes.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.attributes[0],
-                    attr("action", "cancel_ask")
-                );
-                assert_eq!(
-                    cancel_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-
-                assert_eq!(cancel_ask_response.messages.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.messages[0].msg,
-                    transfer_marker_coins(
-                        100,
-                        "con_base_1",
-                        asker_info.sender,
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-                assert_eq!(
-                    cancel_ask_response.messages[1].msg,
-                    transfer_marker_coins(
-                        100,
-                        "base_1",
-                        Addr::unchecked("approver_1"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_ask_invalid_data() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let asker_info = mock_info("asker", &[]);
-
-        // cancel ask order with missing id returns ContractError::Unauthorized
-        let cancel_ask_msg = ExecuteMsg::CancelAsk { id: "".to_string() };
-        let cancel_response = execute(deps.as_mut(), mock_env(), asker_info, cancel_ask_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"id".into()))
-                }
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_ask_non_exist() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let asker_info = mock_info("asker", &[]);
-
-        // cancel non-existent ask order returns ContractError::Unauthorized
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), asker_info, cancel_ask_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::LoadOrderFailed { .. } => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_ask_sender_notequal() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let asker_info = mock_info("asker", &[]);
-
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("not_asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(200),
-            },
-        );
-
-        // cancel ask order with sender not equal to owner returns ContractError::Unauthorized
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), asker_info, cancel_ask_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::Unauthorized => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_ask_with_sent_funds() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // cancel ask order with sent_funds returns ContractError::CancelWithFunds
-        let asker_info = mock_info("asker", &coins(1, "sent_coin"));
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), asker_info, cancel_ask_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::CancelWithFunds => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_bid_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // cancel bid order
-        let bidder_info = mock_info("bidder", &[]);
-
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_bid_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            bidder_info.clone(),
-            cancel_bid_msg,
-        );
-
-        match cancel_bid_response {
-            Ok(cancel_bid_response) => {
-                assert_eq!(cancel_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    cancel_bid_response.attributes[0],
-                    attr("action", "cancel_bid")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(cancel_bid_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: bidder_info.sender.to_string(),
-                        amount: coins(200, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_bid_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"quote_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"quote_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let test_marker: Marker = from_binary(&Binary::from(marker_json)).unwrap();
-        deps.querier.with_markers(vec![test_marker]);
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // cancel bid order
-        let bidder_info = mock_info("bidder", &[]);
-
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_bid_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_bid_response {
-            Ok(cancel_bid_response) => {
-                assert_eq!(cancel_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    cancel_bid_response.attributes[0],
-                    attr("action", "cancel_bid")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-
-                assert_eq!(cancel_bid_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_bid_response.messages[0].msg,
-                    transfer_marker_coins(
-                        200,
-                        "quote_1",
-                        Addr::unchecked("bidder"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_bid_with_fees_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.1".to_string(),
-                }),
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: Some(Coin {
-                    denom: "quote_1".to_string(),
-                    amount: Uint128::new(20),
-                }),
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // cancel bid order
-        let bidder_info = mock_info("bidder", &[]);
-
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_bid_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            bidder_info.clone(),
-            cancel_bid_msg,
-        );
-
-        match cancel_bid_response {
-            Ok(cancel_bid_response) => {
-                assert_eq!(cancel_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    cancel_bid_response.attributes[0],
-                    attr("action", "cancel_bid")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(cancel_bid_response.messages.len(), 2);
-                assert_eq!(
-                    cancel_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: bidder_info.sender.to_string(),
-                        amount: coins(200, "quote_1"),
-                    })
-                );
-                assert_eq!(
-                    cancel_bid_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: bidder_info.sender.to_string(),
-                        amount: coins(20, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_bid_restricted_marker_with_fees() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.1".to_string(),
-                }),
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        let marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"quote_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"quote_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let test_marker: Marker = from_binary(&Binary::from(marker_json)).unwrap();
-        deps.querier.with_markers(vec![test_marker]);
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: Some(Coin {
-                    amount: Uint128::new(20),
-                    denom: "quote_1".to_string(),
-                }),
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // cancel bid order
-        let bidder_info = mock_info("bidder", &[]);
-
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_bid_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_bid_response {
-            Ok(cancel_bid_response) => {
-                assert_eq!(cancel_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    cancel_bid_response.attributes[0],
-                    attr("action", "cancel_bid")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-
-                assert_eq!(cancel_bid_response.messages.len(), 2);
-                assert_eq!(
-                    cancel_bid_response.messages[0].msg,
-                    transfer_marker_coins(
-                        200,
-                        "quote_1",
-                        Addr::unchecked("bidder"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-                assert_eq!(
-                    cancel_bid_response.messages[1].msg,
-                    transfer_marker_coins(
-                        20,
-                        "quote_1",
-                        Addr::unchecked("bidder"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn cancel_bid_invalid_data() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let bidder_info = mock_info("bidder", &[]);
-
-        // cancel bid order with missing id returns ContractError::Unauthorized
-        let cancel_bid_msg = ExecuteMsg::CancelAsk { id: "".to_string() };
-        let cancel_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"id".into()))
-                }
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_bid_non_exist() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let bidder_info = mock_info("bidder", &[]);
-
-        // cancel non-existent bid order returns ContractError::Unauthorized
-        let cancel_bid_msg = ExecuteMsg::CancelAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::LoadOrderFailed { .. } => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_bid_sender_notequal() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let bidder_info = mock_info("bidder", &[]);
-
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(200),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("not_bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // cancel bid order with sender not equal to owner returns ContractError::Unauthorized
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::Unauthorized => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_bid_with_sent_funds() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // cancel bid order with sent_funds returns ContractError::CancelWithFunds
-        let bidder_info = mock_info("bidder", &coins(1, "sent_coin"));
-        let cancel_bid_msg = ExecuteMsg::CancelAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let cancel_response = execute(deps.as_mut(), mock_env(), bidder_info, cancel_bid_msg);
-
-        match cancel_response {
-            Err(error) => match error {
-                ContractError::CancelWithFunds => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn cancel_bid_partial() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(10),
-            },
-        );
-
-        let bid_id = "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b";
-        let ask_id = "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367";
-
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: bid_id.into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: ask_id.into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(10),
-            },
-        );
-
-        let execute_msg = ExecuteMsg::ExecuteMatch {
-            ask_id: ask_id.into(),
-            bid_id: bid_id.into(),
-            price: "2".into(),
-            size: Uint128::new(10),
-        };
-
-        let execute_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            mock_info("exec_1", &[]),
-            execute_msg,
-        );
-
-        match execute_response {
-            Err(error) => panic!("unexpected error: {:?}", error),
-            Ok(..) => {}
-        }
-
-        // cancel bid order
-        let bidder_info = mock_info("bidder", &[]);
-
-        let cancel_bid_msg = ExecuteMsg::CancelBid {
-            id: bid_id.to_string(),
-        };
-
-        let cancel_bid_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            bidder_info.clone(),
-            cancel_bid_msg,
-        );
-
-        match cancel_bid_response {
-            Ok(cancel_bid_response) => {
-                assert_eq!(cancel_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    cancel_bid_response.attributes[0],
-                    attr("action", "cancel_bid")
-                );
-                assert_eq!(cancel_bid_response.attributes[1], attr("id", bid_id));
-                assert_eq!(
-                    cancel_bid_response.attributes[2],
-                    attr("reverse_size", "90")
-                );
-                assert_eq!(
-                    cancel_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(cancel_bid_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: bidder_info.sender.to_string(),
-                        amount: coins(180, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage.load(bid_id.as_bytes()).is_err());
-    }
-
-    #[test]
-    fn cancel_ask_partial() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let bid_id = "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b";
-        let ask_id = "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367";
-
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: ask_id.into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(10),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: bid_id.into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(20),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        let execute_msg = ExecuteMsg::ExecuteMatch {
-            ask_id: ask_id.into(),
-            bid_id: bid_id.into(),
-            price: "2".into(),
-            size: Uint128::new(10),
-        };
-
-        let execute_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            mock_info("exec_1", &[]),
-            execute_msg,
-        );
-
-        match execute_response {
-            Err(error) => panic!("unexpected error: {:?}", error),
-            Ok(..) => {}
-        }
-
-        // cancel ask order
-        let asker_info = mock_info("asker", &[]);
-
-        let cancel_ask_msg = ExecuteMsg::CancelAsk {
-            id: ask_id.to_string(),
-        };
-        let cancel_ask_response = execute(
-            deps.as_mut(),
-            mock_env(),
-            asker_info.clone(),
-            cancel_ask_msg,
-        );
-
-        match cancel_ask_response {
-            Ok(cancel_ask_response) => {
-                assert_eq!(cancel_ask_response.attributes.len(), 2);
-                assert_eq!(
-                    cancel_ask_response.attributes[0],
-                    attr("action", "cancel_ask")
-                );
-                assert_eq!(cancel_ask_response.attributes[1], attr("id", ask_id));
-                assert_eq!(cancel_ask_response.messages.len(), 1);
-                assert_eq!(
-                    cancel_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: asker_info.sender.to_string(),
-                        amount: coins(90, "base_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage.load(ask_id.as_bytes()).is_err());
-    }
-
-    #[test]
-    fn expire_ask_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let expire_ask_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_ask_response {
-            Ok(expire_ask_response) => {
-                assert_eq!(expire_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_ask_response.attributes[0],
-                    attr("action", "expire_ask")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_ask_response.messages.len(), 1);
-                assert_eq!(
-                    expire_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "asker".to_string(),
-                        amount: coins(100, "base_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn reject_ask_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_ask_msg = ExecuteMsg::RejectAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-            size: None,
-        };
-        let reject_ask_response = execute(deps.as_mut(), mock_env(), exec_info, reject_ask_msg);
-
-        match reject_ask_response {
-            Ok(reject_ask_response) => {
-                assert_eq!(reject_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_ask_response.attributes[0],
-                    attr("action", "reject_ask")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(reject_ask_response.messages.len(), 1);
-                assert_eq!(
-                    reject_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "asker".to_string(),
-                        amount: coins(100, "base_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn reject_partial_ask_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(200),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_ask_msg = ExecuteMsg::RejectAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-            size: Some(Uint128::new(100)),
-        };
-        let reject_ask_response = execute(deps.as_mut(), mock_env(), exec_info, reject_ask_msg);
-
-        match reject_ask_response {
-            Ok(reject_ask_response) => {
-                assert_eq!(reject_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_ask_response.attributes[0],
-                    attr("action", "reject_ask")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    reject_ask_response.attributes[3],
-                    attr("order_open", "true")
-                );
-                assert_eq!(reject_ask_response.messages.len(), 1);
-                assert_eq!(
-                    reject_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "asker".to_string(),
-                        amount: coins(100, "base_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order updated
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        match ask_storage.load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    AskOrderV1 {
-                        base: "base_1".into(),
-                        class: AskOrderClass::Basic,
-                        id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                        owner: Addr::unchecked("asker"),
-                        price: "2".into(),
-                        quote: "quote_1".into(),
-                        size: Uint128::new(100)
-                    }
-                )
-            }
-            _ => {
-                panic!("ask order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_ask_cancel_size_not_increment() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_ask_msg = ExecuteMsg::RejectAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-            size: Some(Uint128::new(50)),
-        };
-        let reject_ask_response = execute(deps.as_mut(), mock_env(), exec_info, reject_ask_msg);
-
-        match reject_ask_response {
-            Ok(_) => panic!("expected error, but ok"),
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"size".into()))
-                }
-                error => panic!("unexpected error: {:?}", error),
-            },
-        }
-
-        // verify ask order unchanged
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        match ask_storage.load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    AskOrderV1 {
-                        base: "base_1".into(),
-                        class: AskOrderClass::Basic,
-                        id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                        owner: Addr::unchecked("asker"),
-                        price: "2".into(),
-                        quote: "quote_1".into(),
-                        size: Uint128::new(100)
-                    }
-                )
-            }
-            _ => {
-                panic!("ask order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_ask_cancel_size_greater_than_order_size() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_ask_msg = ExecuteMsg::RejectAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-            size: Some(Uint128::new(150)),
-        };
-        let reject_ask_response = execute(deps.as_mut(), mock_env(), exec_info, reject_ask_msg);
-
-        match reject_ask_response {
-            Ok(_) => panic!("expected error, but ok"),
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"size".into()))
-                }
-                error => panic!("unexpected error: {:?}", error),
-            },
-        }
-
-        // verify ask order unchanged
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        match ask_storage.load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    AskOrderV1 {
-                        base: "base_1".into(),
-                        class: AskOrderClass::Basic,
-                        id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                        owner: Addr::unchecked("asker"),
-                        price: "2".into(),
-                        quote: "quote_1".into(),
-                        size: Uint128::new(100)
-                    }
-                )
-            }
-            _ => {
-                panic!("ask order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn expire_ask_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let test_marker: Marker = from_binary(&Binary::from(marker_json)).unwrap();
-        deps.querier.with_markers(vec![test_marker]);
-
-        // create bid data
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("asker"),
-                class: AskOrderClass::Basic,
-                base: "base_1".into(),
-                quote: "quote_1".into(),
-                price: "2".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_ask_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_ask_response {
-            Ok(expire_ask_response) => {
-                assert_eq!(expire_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_ask_response.attributes[0],
-                    attr("action", "expire_ask")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_ask_response.messages.len(), 1);
-                assert_eq!(
-                    expire_ask_response.messages[0].msg,
-                    transfer_marker_coins(
-                        100,
-                        "base_1",
-                        Addr::unchecked("asker"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn expire_ask_convertible_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("approver_1"), Addr::unchecked("approver_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "con_base_1".into(),
-                class: AskOrderClass::Convertible {
-                    status: AskOrderStatus::Ready {
-                        approver: Addr::unchecked("approver_1"),
-                        converted_base: coin(100, "base_denom"),
-                    },
-                },
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let expire_ask_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_ask_response {
-            Ok(expire_ask_response) => {
-                assert_eq!(expire_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_ask_response.attributes[0],
-                    attr("action", "expire_ask")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_ask_response.messages.len(), 2);
-                assert_eq!(
-                    expire_ask_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "asker".to_string(),
-                        amount: coins(100, "con_base_1"),
-                    })
-                );
-                assert_eq!(
-                    expire_ask_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "approver_1".to_string(),
-                        amount: coins(100, "base_denom"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn expire_ask_convertible_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("approver_1"), Addr::unchecked("approver_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let convertible_marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"con_base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"con_base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let base_marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"base_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"base_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let base_marker: Marker = from_binary(&Binary::from(base_marker_json)).unwrap();
-        let convertible_marker: Marker =
-            from_binary(&Binary::from(convertible_marker_json)).unwrap();
-        deps.querier
-            .with_markers(vec![base_marker, convertible_marker]);
-
-        // store valid ask order
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "con_base_1".into(),
-                class: AskOrderClass::Convertible {
-                    status: AskOrderStatus::Ready {
-                        approver: Addr::unchecked("approver_1"),
-                        converted_base: coin(100, "base_1"),
-                    },
-                },
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(100),
-            },
-        );
-
-        // expire ask order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-        let expire_ask_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_ask_response {
-            Ok(expire_ask_response) => {
-                assert_eq!(expire_ask_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_ask_response.attributes[0],
-                    attr("action", "expire_ask")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[1],
-                    attr("id", "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_ask_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_ask_response.messages.len(), 2);
-                assert_eq!(
-                    expire_ask_response.messages[0].msg,
-                    transfer_marker_coins(
-                        100,
-                        "con_base_1",
-                        Addr::unchecked("asker"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-                assert_eq!(
-                    expire_ask_response.messages[1].msg,
-                    transfer_marker_coins(
-                        100,
-                        "base_1",
-                        Addr::unchecked("approver_1"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify ask order removed from storage
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn expire_ask_invalid_data() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("exec_1", &[]);
-
-        // expire ask order with missing id returns ContractError::Unauthorized
-        let expire_ask_msg = ExecuteMsg::ExpireAsk { id: "".to_string() };
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"id".into()))
-                }
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_ask_non_exist() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("exec_1", &[]);
-
-        // expire non-existent ask order returns ContractError::Unauthorized
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::LoadOrderFailed { .. } => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_ask_sender_notequal() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("not_exec", &[]);
-
-        store_test_ask(
-            &mut deps.storage,
-            &AskOrderV1 {
-                base: "base_1".into(),
-                class: AskOrderClass::Basic,
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-                owner: Addr::unchecked("asker"),
-                price: "2".into(),
-                quote: "quote_1".into(),
-                size: Uint128::new(200),
-            },
-        );
-
-        // expire ask order with sender not equal to owner returns ContractError::Unauthorized
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::Unauthorized => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_ask_with_sent_funds() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // expire ask order with sent_funds returns ContractError::ExpireWithFunds
-        let exec_info = mock_info("exec_1", &coins(1, "sent_coin"));
-        let expire_ask_msg = ExecuteMsg::ExpireAsk {
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_ask_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::ExpireWithFunds => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_bid_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                price: "2".into(),
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_bid_msg = ExecuteMsg::ExpireBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_bid_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_bid_response {
-            Ok(expire_bid_response) => {
-                assert_eq!(expire_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_bid_response.attributes[0],
-                    attr("action", "expire_bid")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_bid_response.messages.len(), 1);
-                assert_eq!(
-                    expire_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(200, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn reject_bid_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: None,
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "reject_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 1);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(200, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn reject_partial_bid_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(200),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(400),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(100)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "reject_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "true")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 1);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(200, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order update
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        owner: Addr::unchecked("bidder"),
-                        base: Coin {
-                            amount: Uint128::new(200),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![Event {
-                            action: Action::Reject {
-                                base: Coin {
-                                    denom: "base_1".to_string(),
-                                    amount: Uint128::new(100)
-                                },
-                                fee: None,
-                                quote: Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(200)
-                                },
-                            },
-                            block_info: mock_env().block.into(),
-                        }],
-                        fee: None,
-                        quote: Coin {
-                            amount: Uint128::new(400),
-                            denom: "quote_1".into(),
-                        },
-                        price: "2".into(),
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_bid_with_fees_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.1".to_string(),
-                }),
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: Some(Coin {
-                    amount: Uint128::new(10),
-                    denom: "quote_1".to_string(),
-                }),
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-                price: "1".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(50)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "reject_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "50")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "true")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 2);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(50, "quote_1"),
-                    })
-                );
-                assert_eq!(
-                    reject_bid_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(5, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order update
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        owner: Addr::unchecked("bidder"),
-                        base: Coin {
-                            amount: Uint128::new(100),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![Event {
-                            action: Action::Reject {
-                                base: Coin {
-                                    denom: "base_1".to_string(),
-                                    amount: Uint128::new(50)
-                                },
-                                fee: Some(Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(5)
-                                }),
-                                quote: Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(50)
-                                },
-                            },
-                            block_info: mock_env().block.into(),
-                        }],
-                        fee: Some(Coin {
-                            amount: Uint128::new(10),
-                            denom: "quote_1".to_string(),
-                        }),
-                        quote: Coin {
-                            amount: Uint128::new(100),
-                            denom: "quote_1".into(),
-                        },
-                        price: "1".into(),
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_bid_with_fees_round_down() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.1".to_string(),
-                }),
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: Some(Coin {
-                    amount: Uint128::new(10),
-                    denom: "quote_1".to_string(),
-                }),
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-                price: "1".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(55)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "reject_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "55")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "true")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 2);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(55, "quote_1"),
-                    })
-                );
-                assert_eq!(
-                    reject_bid_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(5, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order update
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        owner: Addr::unchecked("bidder"),
-                        base: Coin {
-                            amount: Uint128::new(100),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![Event {
-                            action: Action::Reject {
-                                base: Coin {
-                                    denom: "base_1".to_string(),
-                                    amount: Uint128::new(55)
-                                },
-                                fee: Some(Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(5)
-                                }),
-                                quote: Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(55)
-                                },
-                            },
-                            block_info: mock_env().block.into(),
-                        }],
-                        fee: Some(Coin {
-                            amount: Uint128::new(10),
-                            denom: "quote_1".to_string(),
-                        }),
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        quote: Coin {
-                            amount: Uint128::new(100),
-                            denom: "quote_1".into(),
-                        },
-                        price: "1".into(),
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_bid_with_fees_round_up() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.1".to_string(),
-                }),
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: Some(Coin {
-                    amount: Uint128::new(10),
-                    denom: "quote_1".to_string(),
-                }),
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-                price: "1".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(56)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "reject_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "56")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "true")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 2);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(56, "quote_1"),
-                    })
-                );
-                assert_eq!(
-                    reject_bid_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(6, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order update
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        owner: Addr::unchecked("bidder"),
-                        base: Coin {
-                            amount: Uint128::new(100),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![Event {
-                            action: Action::Reject {
-                                base: Coin {
-                                    denom: "base_1".to_string(),
-                                    amount: Uint128::new(56)
-                                },
-                                fee: Some(Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(6)
-                                }),
-                                quote: Coin {
-                                    denom: "quote_1".to_string(),
-                                    amount: Uint128::new(56)
-                                },
-                            },
-                            block_info: mock_env().block.into(),
-                        }],
-                        fee: Some(Coin {
-                            amount: Uint128::new(10),
-                            denom: "quote_1".to_string(),
-                        }),
-                        quote: Coin {
-                            amount: Uint128::new(100),
-                            denom: "quote_1".into(),
-                        },
-                        price: "1".into(),
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_partial_bid_cancel_size_not_increment() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(50)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(_) => panic!("expected error, but ok"),
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"size".into()))
-                }
-                error => panic!("unexpected error: {:?}", error),
-            },
-        }
-
-        // verify bid order not updated
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        base: Coin {
-                            amount: Uint128::new(100),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![],
-                        fee: None,
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        owner: Addr::unchecked("bidder"),
-                        price: "2".into(),
-                        quote: Coin {
-                            amount: Uint128::new(200),
-                            denom: "quote_1".into(),
-                        },
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn reject_bid_cancel_size_greater_than_order_size() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let reject_bid_msg = ExecuteMsg::RejectBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-            size: Some(Uint128::new(150)),
-        };
-
-        let reject_bid_response = execute(deps.as_mut(), mock_env(), exec_info, reject_bid_msg);
-
-        match reject_bid_response {
-            Ok(_) => panic!("expected error, but ok"),
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"size".into()))
-                }
-                error => panic!("unexpected error: {:?}", error),
-            },
-        }
-
-        // verify bid order not updated
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        match bid_storage.load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes()) {
-            Ok(stored_order) => {
-                assert_eq!(
-                    stored_order,
-                    BidOrderV2 {
-                        base: Coin {
-                            amount: Uint128::new(100),
-                            denom: "base_1".into(),
-                        },
-                        events: vec![],
-                        fee: None,
-                        id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                        owner: Addr::unchecked("bidder"),
-                        price: "2".into(),
-                        quote: Coin {
-                            amount: Uint128::new(200),
-                            denom: "quote_1".into(),
-                        },
-                    }
-                )
-            }
-            _ => {
-                panic!("bid order was not found in storage")
-            }
-        }
-    }
-
-    #[test]
-    fn expire_bid_restricted_marker() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let marker_json = b"{
-              \"address\": \"tp18vmzryrvwaeykmdtu6cfrz5sau3dhc5c73ms0u\",
-              \"coins\": [
-                {
-                  \"denom\": \"quote_1\",
-                  \"amount\": \"1000\"
-                }
-              ],
-              \"account_number\": 10,
-              \"sequence\": 0,
-              \"permissions\": [
-                {
-                  \"permissions\": [
-                    \"burn\",
-                    \"delete\",
-                    \"deposit\",
-                    \"admin\",
-                    \"mint\",
-                    \"withdraw\"
-                  ],
-                  \"address\": \"tp18vd8fpwxzck93qlwghaj6arh4p7c5n89x8kskz\"
-                }
-              ],
-              \"status\": \"active\",
-              \"denom\": \"quote_1\",
-              \"total_supply\": \"1000\",
-              \"marker_type\": \"restricted\",
-              \"supply_fixed\": false
-            }";
-
-        let test_marker: Marker = from_binary(&Binary::from(marker_json)).unwrap();
-        deps.querier.with_markers(vec![test_marker]);
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(100),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(200),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_bid_msg = ExecuteMsg::ExpireBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_bid_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_bid_response {
-            Ok(expire_bid_response) => {
-                assert_eq!(expire_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    expire_bid_response.attributes[0],
-                    attr("action", "expire_bid")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[2],
-                    attr("reverse_size", "100")
-                );
-                assert_eq!(
-                    expire_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(expire_bid_response.messages.len(), 1);
-                assert_eq!(
-                    expire_bid_response.messages[0].msg,
-                    transfer_marker_coins(
-                        200,
-                        "quote_1",
-                        Addr::unchecked("bidder"),
-                        Addr::unchecked(MOCK_CONTRACT_ADDR)
-                    )
-                    .unwrap()
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-
-        // verify bid order removed from storage
-        let bid_storage = get_bid_storage_read(&deps.storage);
-        assert!(bid_storage
-            .load("c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".as_bytes())
-            .is_err());
-    }
-
-    #[test]
-    fn expire_bid_invalid_data() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("exec_1", &[]);
-
-        // expire bid order with missing id returns ContractError::Unauthorized
-        let expire_bid_msg = ExecuteMsg::ExpireAsk { id: "".to_string() };
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::InvalidFields { fields } => {
-                    assert!(fields.contains(&"id".into()))
-                }
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_bid_non_exist() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("exec_1", &[]);
-
-        // expire non-existent bid order returns ContractError::Unauthorized
-        let expire_bid_msg = ExecuteMsg::ExpireAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::LoadOrderFailed { .. } => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_bid_sender_notequal() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        let exec_info = mock_info("not_exec", &[]);
-
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(200),
-                    denom: "base_1".into(),
-                },
-                events: vec![],
-                fee: None,
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-                price: "2".into(),
-            },
-        );
-
-        // expire bid order with sender not equal to owner returns ContractError::Unauthorized
-        let expire_bid_msg = ExecuteMsg::ExpireBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::Unauthorized => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_bid_with_sent_funds() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // expire bid order with sent_funds returns ContractError::ExpireWithFunds
-        let exec_info = mock_info("exec_1", &coins(1, "sent_coin"));
-        let expire_bid_msg = ExecuteMsg::ExpireAsk {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_response {
-            Err(error) => match error {
-                ContractError::ExpireWithFunds => {}
-                _ => {
-                    panic!("unexpected error: {:?}", error)
-                }
-            },
-            Ok(_) => panic!("expected error, but ok"),
-        }
-    }
-
-    #[test]
-    fn expire_partial_filled_bid_with_fees_valid() {
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: Some(FeeInfo {
-                    account: Addr::unchecked("bid_fee_account"),
-                    rate: "0.003".to_string(),
-                }),
-                ask_required_attributes: vec![],
-                bid_required_attributes: vec![],
-                price_precision: Uint128::new(0),
-                size_increment: Uint128::new(1),
-            },
-        );
-
-        // create bid data
-        store_test_bid(
-            &mut deps.storage,
-            &BidOrderV2 {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-                owner: Addr::unchecked("bidder"),
-                base: Coin {
-                    amount: Uint128::new(10000),
-                    denom: "base_1".into(),
-                },
-                events: vec![Event {
-                    action: Action::Fill {
-                        base: Coin {
-                            denom: "base_1".to_string(),
-                            amount: Uint128::new(2000),
-                        },
-                        fee: Some(Coin {
-                            denom: "quote_1".to_string(),
-                            amount: Uint128::new(6),
-                        }),
-                        price: "0.01".to_string(),
-                        quote: Coin {
-                            denom: "quote_1".to_string(),
-                            amount: Uint128::new(20),
-                        },
-                    },
-                    block_info: mock_env().block.into(),
-                }],
-                fee: Some(Coin {
-                    amount: Uint128::new(30),
-                    denom: "quote_1".to_string(),
-                }),
-                quote: Coin {
-                    amount: Uint128::new(100),
-                    denom: "quote_1".into(),
-                },
-                price: "0.01".into(),
-            },
-        );
-
-        // expire bid order
-        let exec_info = mock_info("exec_1", &[]);
-
-        let expire_bid_msg = ExecuteMsg::ExpireBid {
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".to_string(),
-        };
-
-        let expire_bid_response = execute(deps.as_mut(), mock_env(), exec_info, expire_bid_msg);
-
-        match expire_bid_response {
-            Ok(reject_bid_response) => {
-                assert_eq!(reject_bid_response.attributes.len(), 4);
-                assert_eq!(
-                    reject_bid_response.attributes[0],
-                    attr("action", "expire_bid")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[1],
-                    attr("id", "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[2],
-                    attr("reverse_size", "8000")
-                );
-                assert_eq!(
-                    reject_bid_response.attributes[3],
-                    attr("order_open", "false")
-                );
-                assert_eq!(reject_bid_response.messages.len(), 2);
-                assert_eq!(
-                    reject_bid_response.messages[0].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(80, "quote_1"),
-                    })
-                );
-                assert_eq!(
-                    reject_bid_response.messages[1].msg,
-                    CosmosMsg::Bank(BankMsg::Send {
-                        to_address: "bidder".to_string(),
-                        amount: coins(24, "quote_1"),
-                    })
-                );
-            }
-            Err(error) => panic!("unexpected error: {:?}", error),
-        }
-    }
-
-    #[test]
     fn execute_valid_data() {
         // setup
         let mut deps = mock_dependencies(&[]);
@@ -14685,123 +10850,6 @@ mod tests {
             }
             Err(error) => panic!("unexpected error: {:?}", error),
         }
-    }
-
-    #[test]
-    fn query_ask_order() {
-        // setup
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid ask order
-        let ask_order = AskOrderV1 {
-            base: "base_1".into(),
-            class: AskOrderClass::Basic,
-            id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-            owner: Addr::unchecked("asker"),
-            price: "2".into(),
-            quote: "quote_1".into(),
-            size: Uint128::new(200),
-        };
-
-        let mut ask_storage = get_ask_storage(&mut deps.storage);
-        if let Err(error) = ask_storage.save(
-            "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes(),
-            &ask_order,
-        ) {
-            panic!("unexpected error: {:?}", error)
-        };
-
-        // verify ask order still exists
-        let ask_storage = get_ask_storage_read(&deps.storage);
-        assert!(ask_storage
-            .load("ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".as_bytes())
-            .is_ok());
-
-        // query for ask order
-        let query_ask_response = query(
-            deps.as_ref(),
-            mock_env(),
-            QueryMsg::GetAsk {
-                id: "ab5f5a62-f6fc-46d1-aa84-51ccc51ec367".into(),
-            },
-        );
-
-        assert_eq!(query_ask_response, to_binary(&ask_order));
-    }
-
-    #[test]
-    fn query_bid_order() {
-        // setup
-        let mut deps = mock_dependencies(&[]);
-        setup_test_base(
-            &mut deps.storage,
-            &ContractInfoV3 {
-                name: "contract_name".into(),
-                bind_name: "contract_bind_name".into(),
-                base_denom: "base_denom".into(),
-                convertible_base_denoms: vec!["con_base_1".into(), "con_base_2".into()],
-                supported_quote_denoms: vec!["quote_1".into(), "quote_2".into()],
-                approvers: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                executors: vec![Addr::unchecked("exec_1"), Addr::unchecked("exec_2")],
-                ask_fee_info: None,
-                bid_fee_info: None,
-                ask_required_attributes: vec!["ask_tag_1".into(), "ask_tag_2".into()],
-                bid_required_attributes: vec!["bid_tag_1".into(), "bid_tag_2".into()],
-                price_precision: Uint128::new(2),
-                size_increment: Uint128::new(100),
-            },
-        );
-
-        // store valid bid order
-        let bid_order = BidOrderV2 {
-            base: Coin {
-                amount: Uint128::new(100),
-                denom: "base_1".into(),
-            },
-            events: vec![],
-            fee: None,
-            id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-            owner: Addr::unchecked("bidder"),
-            price: "2".into(),
-            quote: Coin {
-                amount: Uint128::new(100),
-                denom: "quote_1".into(),
-            },
-        };
-
-        let mut bid_storage = get_bid_storage(&mut deps.storage);
-        if let Err(error) = bid_storage.save(bid_order.id.as_bytes(), &bid_order) {
-            panic!("unexpected error: {:?}", error);
-        };
-
-        // query for bid order
-        let query_bid_response = query(
-            deps.as_ref(),
-            mock_env(),
-            QueryMsg::GetBid {
-                id: "c13f8888-ca43-4a64-ab1b-1ca8d60aa49b".into(),
-            },
-        );
-
-        assert_eq!(query_bid_response, to_binary(&bid_order));
     }
 
     fn setup_test_base(storage: &mut dyn Storage, contract_info: &ContractInfoV3) {
